+++ src/eduid_webapp/signup/templates/verification_email.html.jinja2
@@ -1,0 +1,11 @@
+{% trans verification_link=verification_link, site_url=site_url, site_name=site_name %}
+
+<h2>Welcome to {{ site_name }},</h2>
+
+<p>You recently signed up for <a href="{{ site_url }}">{{ site_name }}</a>.</p>
+
+<p>Please confirm the e-mail address and get your password by clicking on this link:</p>
+
+<a href="{{ verification_link }}">{{ verification_link }}</a>
+
+{% endtrans %}
