+++ src/eduid_webapp/signup/templates/verification_email.txt.jinja2
@@ -1,0 +1,11 @@
+{% trans verification_link=verification_link, site_url=site_url, site_name=site_name %}
+
+Welcome to {{ site_name }},
+
+You recently signed up for {{ site_name }}.
+
+Please confirm the e-mail address and get your password by clicking on this link:
+
+  {{ verification_link }}
+
+{% endtrans %}
